chore: polish migrate workflow
