+++ modular/gater/router_test.go
@@ -51,6 +51,14 @@ func TestRouters(t *testing.T) {
 			name:             "Get create object approval router",
 			router:           gwRouter,
 			method:           http.MethodGet,
+			url:              scheme + testDomain + MigratePiecePath,
+			shouldMatch:      true,
+			wantedRouterName: migratePieceRouterName,
+		},
+		{
+			name:             "Migrate piece data",
+			router:           gwRouter,
+			method:           http.MethodGet,
 			url:              scheme + testDomain + GetApprovalPath + "?" + ActionQuery + "=" + createObjectApprovalAction,
 			shouldMatch:      true,
 			wantedRouterName: approvalRouterName,
